Dolphin True Crime New York City Edition (TM)
Replicates the behavior of True Crime New York City's floating point exception handler directly in Dolphin. This is an awful hack to establish a baseline for how the game should behave under Dolphin. I have an actual implementation of floating point exceptions in the works, and this gives me something to compare against.

Only works on Windows on x64. Both JIT and the interpreter are supported.

This change also contains some logging to make it very obvious when a game enables floating point exceptions. Enable the PowerPC log type to see it. For example, True Crime will cause the following log lines to be printed:

[PowerPC]: MSR[FE0] set
[PowerPC]: MSR[FE1] set
[PowerPC]: FPSCR[ZE] set